+++ frontend/reporting/src/app/areas/_help/templates/tab-examples.ts
@@ -10,11 +10,11 @@ export const tabExamplesTemplate = `<ng-template #tabExamplesTemplate>
     <em>DocumentBurster</em> as a general
     <a
       title="Examples of how people are using DocumentBurster report distribution software"
-      href="https://www.reportburster.com/blog/report-distribution-software/"
+      href="https://www.pdfburst.com/blog/report-distribution-software/"
       >Report Distribution Software</a
     >&nbsp;&nbsp;
     <a
-      href="https://www.reportburster.com/blog/report-distribution-software/"
+      href="https://www.pdfburst.com/blog/report-distribution-software/"
       target="_blank"
     >
       <button class="btn btn-primary btn-xs" type="button">
@@ -28,12 +28,12 @@ export const tabExamplesTemplate = `<ng-template #tabExamplesTemplate>
 
     <a
       title="Examples of how people are using DocumentBurster to burst and distribute Crystal Reports documents"
-      href="https://www.reportburster.com/blog/crystal-reports-distribution/"
+      href="https://www.pdfburst.com/blog/crystal-reports-distribution/"
       >Burst and Distribute Crystal Reports Documents</a
     >
     using <em>DocumentBurster</em>&nbsp;&nbsp;
     <a
-      href="https://www.reportburster.com/blog/crystal-reports-distribution/"
+      href="https://www.pdfburst.com/blog/crystal-reports-distribution/"
       target="_blank"
     >
       <button class="btn btn-primary btn-xs" type="button">
@@ -46,11 +46,11 @@ export const tabExamplesTemplate = `<ng-template #tabExamplesTemplate>
 
     <a
       title="Examples of how people are using DocumentBurster for emailing payslips to employees"
-      href="https://www.reportburster.com/blog/email-payslips/"
+      href="https://www.pdfburst.com/blog/email-payslips/"
       >Email Payslips to Employees</a
     >
     using <em>DocumentBurster</em>&nbsp;&nbsp;
-    <a href="https://www.reportburster.com/blog/email-payslips/" target="_blank">
+    <a href="https://www.pdfburst.com/blog/email-payslips/" target="_blank">
       <button class="btn btn-primary btn-xs" type="button">
         View Examples
       </button>