+++ frontend/reporting/src/app/areas/install-setup-upgrade/extra-packages/extra-packages.template.ts
@@ -1,14 +1,7 @@
 export const extraPackagesTemplate = `<!-- <ng-template #extraPackagesTemplate> -->
 
   <div *ngIf="bashService.isChocoOk">
-    <h5>
-      <strong
-        >Do more with <em>DocumentBurster</em> by installing more useful 
-        packages</strong
-      >
-    </h5>
-    <br />
-
+    <span [innerHTML]="'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.INNER-HTML.ABOUT' | translate"></span>
     <div
       *ngFor="let extraPackage of this.extraPackages"
       [ngClass]="{'panel panel-primary': extraPackage.status != 'not-installed', 'panel panel-default' : extraPackage.status == 'not-installed' }"
@@ -16,49 +9,27 @@ export const extraPackagesTemplate = `<!-- <ng-template #extraPackagesTemplate> 
       <div class="panel-heading">
         <h3 class="panel-title">
           <a href="{{extraPackage.website}}" target="_blank"
-            >{{extraPackage.name}}
+            ><span>{{extraPackage.website}}</span
+            >&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
             <span *ngIf="extraPackage.status == 'not-installed'"
-              >(<em>not installed</em>)</span
+              ><em>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.INSTALLED.NOT' | translate }}</em></span
             ><span *ngIf="extraPackage.status != 'not-installed'"
-              >(<em>installed</em>)</span
+              ><strong><u>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.INSTALLED.ALREADY' | translate }}</u></strong></span
             ></a
           >
-          &nbsp;
-          <button
-            id="btnInstall"
-            type="button"
-            class="btn btn-xs btn-primary"
-            *ngIf="extraPackage.status === 'not-installed'"
-            (click)="doInstallUninstallAction(extraPackage, 'install')"
-            [disabled]="executionStatsService.jobStats.numberOfActiveJobs > 0"
-          >
-            <strong
-              >{{'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.JAVA.INSTALL' |
-              translate }}&nbsp;{{extraPackage.name}}</strong
-            >
-          </button>
-          <button
-            id="btnUnInstall"
-            type="button"
-            class="btn btn-xs btn-default"
-            *ngIf="extraPackage.status === 'installed'"
-            (click)="doInstallUninstallAction(extraPackage, 'uninstall')"
-            [disabled]="executionStatsService.jobStats.numberOfActiveJobs > 0"
-          >
-            <strong
-              >{{'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.JAVA.UNINSTALL' |
-              translate }}&nbsp;{{extraPackage.name}}</strong
-            >
-          </button>
+          
         </h3>
       </div>
 
-      <div class="panel-body">
+      <div id="package-{{extraPackage.id}}" class="panel-body">
         <a href="{{extraPackage.website}}" target="_blank"
           ><img src="assets/images/{{extraPackage.icon}}" />
           {{extraPackage.name}}</a
         >
         {{extraPackage.description}}
+        <br/><br/>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.COMMANDS.INSTALL' | translate }}:&nbsp;<em><strong>{{extraPackage.cmdInstall}}</strong></em>
+        <br/>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.COMMANDS.UNINSTALL' | translate }}:&nbsp;<em><strong>{{extraPackage.cmdUnInstall}}</strong></em>
+        <br/>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.COMMANDS.GET-INFO' | translate }}:&nbsp;<em><strong>{{extraPackage.cmdGetInfo}}</strong></em>
         <br />
       </div>
     </div>